added windows unlink support
Windows is picky aboit using is_file before unlink. There you go.

+++ uploader_back.php
@@ -295,7 +295,8 @@ class Uploader
         {
             if(file_exists($file))
             {
-                unlink($file);
+                if(is_file($file))
+                    unlink($file);
             }
         }
     }
